refactor: isolate delegated child launch policy
